requirements.txt
pillow==12.1.1
black==26.3.1
pytest==9.0.2